+++ R/write_camtrapdp.R
@@ -27,10 +27,11 @@
 write_camtrapdp <- function(x, directory, ...) {
   check_camtrapdp(x)
 
-  # Remove columns that were added by read_camtrapdp()
+  # Remove columns and duplicates that were added by read_camtrapdp()
   media(x) <-
     media(x) %>%
-    dplyr::select(-"eventID")
+    dplyr::select(-"eventID") %>%
+    dplyr::distinct()
   # Do not use assign function for observations because it would update
   # x$taxonomic without the "taxon."-terms
   x$data$observations <-
@@ -55,6 +56,9 @@ write_camtrapdp <- function(x, directory, ...) {
 
   # Remove data
   x$data <- NULL
+
+  # Remove NA values in x$taxonomic to avoid validation error in frictionless
+  x$taxonomic <- purrr::map(x$taxonomic, ~ .x[!is.na(.x)])
 
   # Write files
   frictionless::write_package(x, directory, ...)